fix: fix req script to use CLI instead of bootstrap.ts script

+++ scripts/install/50-oracle-ctr-sb.sh
@@ -7,16 +7,28 @@ source ./00-vars.cfg
 export host_dir="$(pwd)/../src"
 export ctr_dir="/app"
 
-ctr image pull docker.io/library/node:lts-alpine3.19
 CTR_NAME="CTR-sg-tmp-node"
 set +e
 pkill -15 "${CTR_NAME}"
 set -e
-ctr run -t --net-host \
-	--rm --cwd "${ctr_dir}" \
-	--mount type=bind,src=${host_dir},dst="${ctr_dir}",options=rbind:rw \
-	docker.io/library/node:lts-alpine3.19 \
-	"${CTR_NAME}" /bin/sh
+
+export use_docker="${1:-''}"
+export image="docker.io/library/node:22-bookworm"
+
+if [[ "${use_docker}" != "--docker" ]]; then
+	ctr image pull "${image}"
+	ctr run -t --net-host \
+		--rm --cwd "${ctr_dir}" \
+		--mount type=bind,src=${host_dir},dst="${ctr_dir}",options=rbind:rw \
+		"${image}" "${CTR_NAME}" /bin/bash
+else
+	docker image pull "${image}"
+	docker run -it --rm \
+		--workdir "${ctr_dir}" \
+		--mount type=bind,src=${host_dir},dst="${ctr_dir}" \
+		--name "${CTR_NAME}" --entrypoint /bin/bash \
+		"${image}"
+fi
 
 echo "!!! IMPORTANT !!!"
 echo "The output from last command represents your Oracle/Guardian public keys (and related data)."
